Add genqlient
We're getting close to ready to make genqlient "maintained open-source"!
So it's time to add it here.  In general I think this page needs some
reordering but I just put it in at a seemingly reasonable place to
start.  The header image from its README doesn't really work well at
this size, so I just left it without a logo for now; we can add one
when someone gets a chance to make one.

I'll hold off on landing this until genqlient#96 lands.

Test plan: `npm run build` and view in a browser

Reviewers: jeff, kevinb